fix(contracts): PRD_104 指令集收敛——移除 Merge 成员与求值分支，词表外操作码注册边界显式拒绝
- InstructionOp 词表收敛为 SIGNAL/NOT/AND/OR/DELAY（解冻窗口一并收口）；
  求值器/校验器移除扇入分支与 _mergeValue，锚点/裸 SIGNAL 注释改述为
  NOT/OR 口径（编码层契约保留，仅不再引用已退役指令）
- Instruction.op 改 uint8 数值承载（ABI 签名/EIP-712/hooksHash 编码与
  枚举形态逐字节一致）：词表外操作码（如已退役 op=5）在 commitPlan 注册
  边界被 _validateHook 显式 revert InvalidInstruction，不再依赖解码层的
  无名回滚；求值器补同款词表兜底拒绝
- 测试：移除两条 Merge 用例，新增数值拼装 op=5 的注册边界拒绝用例；
  UVPDockingModule.t 的 op 字面量随 uint8 化补 cast
- hook-core：旧扇入标头拒绝断言改为通用语法错误口径（词元按字节拼装，
  镜像 uvp-core 解析器零命中做法），README 退役清单同步移除该词
- fixtures v0.10 原位重钉（先例 17bd417/52225a9：内部数据变更不换版）：
  仅 4 项 hash 变更，events/functions（selectors/topics）逐字节不变；
  abiHash 变更仅源于 artifact internalType 元数据 enum→uint8（wire 类型
  本就是 uint8）；protocol-bindings artifact-abis 同步重生成（单行
  internalType 差异）

+++ contracts/uvp-contracts/test/UVPDockingModule.t.sol
@@ -915,7 +915,7 @@ contract UVPDockingModuleTest {
         UVPStateMachine.CompactHook[] memory hooks = new UVPStateMachine.CompactHook[](2);
         UVPStateMachine.Instruction[] memory instructions = new UVPStateMachine.Instruction[](1);
         instructions[0] = UVPStateMachine.Instruction({
-            op: UVPStateMachine.InstructionOp.Signal,
+            op: uint8(UVPStateMachine.InstructionOp.Signal),
             sourceId: TARGET_SOURCE,
             signalId: TARGET_SIGNAL,
             arity: 0,
@@ -937,7 +937,7 @@ contract UVPDockingModuleTest {
         // outside 触发的出生事实必须在 plan capability 词表内。
         UVPStateMachine.Instruction[] memory mintInstructions = new UVPStateMachine.Instruction[](1);
         mintInstructions[0] = UVPStateMachine.Instruction({
-            op: UVPStateMachine.InstructionOp.Signal,
+            op: uint8(UVPStateMachine.InstructionOp.Signal),
             sourceId: TARGET_OUT_SOURCE,
             signalId: TARGET_OUT_SIGNAL,
             arity: 0,
@@ -1005,7 +1005,7 @@ contract UVPDockingModuleTest {
     ) private pure returns (UVPStateMachine.CompactHook memory) {
         UVPStateMachine.Instruction[] memory instructions = new UVPStateMachine.Instruction[](1);
         instructions[0] = UVPStateMachine.Instruction({
-            op: UVPStateMachine.InstructionOp.Signal, sourceId: sourceId, signalId: signalId, arity: 0, delaySeconds: 0
+            op: uint8(UVPStateMachine.InstructionOp.Signal), sourceId: sourceId, signalId: signalId, arity: 0, delaySeconds: 0
         });
         bytes32[] memory deps = new bytes32[](1);
         deps[0] = keccak256(abi.encode(sourceId, signalId));
